requirements.txt
# Core dependencies
torch==2.0.1
torchvision==0.15.2
transformers==4.30.2
opencv-python==4.8.1.78
pillow==9.5.0
numpy==1.24.3

# Additional utilities
tqdm==4.65.0
matplotlib==3.7.1
pycocotools==2.0.6
huggingface-hub==0.16.4
datasets==2.12.0

# Optional dependencies for development
pytest==7.3.1
black==23.3.0
flake8==6.0.0 